remove sudo and register cleanup function
Register the cleanup function so that it is always called
on script exit signal, whatever that may be. This ensures we dont
end up exiting in error before we have a chance to clean up the
file with secrets in it.

Remove multipass install in favour of informing the user to
install multipass so that we may remove usage of sudo from script.