feat: make the deploy script environment aware. (#34)
This commit modifies the democluster user-data and the public script for
deploying the democluster in order to make the deployment environment
aware.

The main modification in the image itself is that the snaps are
not installed during the build of the image, instead, they are when
the image is being launched. This allows the snap to be the most up
to date according to the environment.

In the deploy script, it is added support for a third argument that is
the environment, i.e. `dev`, `qa` or `staging`. When it is null, it
defaults to production, which guarantees that end users of Vantage in
production will not get in touch with this third argument.